Refactor DB things... Separate run and query
